[SPARK-54410][SQL] Fix read support for the variant logical type annotation
### What changes were proposed in this pull request?

[This PR](https://github.com/apache/spark/pull/53005) introduced a fix where the Spark parquet writer would annotate variant columns with the parquet variant logical type. The PR had an ad-hoc fix on the reader side for validation. This PR formally allows Spark to read parquet files with the Variant logical type.

The PR also introduces an unrelated fix in ParquetRowConverter to allow Spark to read variant columns regardless of which order the value and metadata fields are stored in.

### Why are the changes needed?

The variant logical type annotation has formally been adopted as part of the parquet spec in is part of the parquet-java 1.16.0 library. Therefore, Spark should be able to read files containing data annotated as such.

### Does this PR introduce _any_ user-facing change?

Yes, it allows users to read parquet files with the variant logical type annotation.

### How was this patch tested?

Existing test from [this PR](https://github.com/apache/spark/pull/53005) where we wrote data of the variant logical type and tested read using an ad-hoc solution.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53120 from harshmotw-db/harshmotw-db/variant_annotation_write.

Authored-by: Harsh Motwani <harsh.motwani@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>