[SPARK-54439][SQL] KeyGroupedPartitioning and join key size mismatch
### What changes were proposed in this pull request?

Fix `KeyGroupedShuffleSpec.createPartitioning()` as clustering required at the other side of the join might contain more clustering expressions than the number of expressions in the shuffle spec's `KeyGroupedPartitioning`, so simply zipping them is not correct.

### Why are the changes needed?

Fix a correctness issue due to wrong partitioning on the shuffle side.

### Does this PR introduce _any_ user-facing change?

Yes, it fixes the query.

### How was this patch tested?

Added new UT.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53142 from peter-toth/SPARK-54439-keygroupedpartitioning-and-join-key-size-mismatch.

Authored-by: Peter Toth <peter.toth@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>