[SPARK-54452] Fix empty response from SparkConnect server for `spark.sql(...)` inside FlowFunction
### What changes were proposed in this pull request?

In PR #53024, we added SDP support for `spark.sql(...)` inside a FlowFunction. For these calls, instead of eagerly executing the SQL, the Spark Connect server should return the raw logical plan to the client and defer execution to the flow function.

However, in that PR we constructed the response object but forgot to actually return it to the Spark Connect client, so the client received an empty response.

This went unnoticed in tests because, when the client sees an empty `spark.sql(...)` response, [it falls back to creating an empty DataFrame holding the raw logical plan](https://github.com/apache/spark/blob/master/python/pyspark/sql/connect/session.py#L829-L835), which happens to match the desired behavior. This PR fixes the bug by returning the proper response instead of relying on that implicit fallback.

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

This PR fixes a bug introduced in #53024 where the server did not return the constructed spark.sql(...) response to the client.

### How was this patch tested?

New tests

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53156 from SCHJonathan/jonathan-chang_data/fix-spark-sql-bug.

Authored-by: Yuheng Chang <jonathanyuheng@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 997525cae9b00e47626ce09f00e779b279551ace)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>