[SPARK-54458][INFRA] Use official download-artifact to fix the test_report action
### What changes were proposed in this pull request?

We should fix the `test_report` workflow - it barely works.

* Used the official download-artifacts to replace the 3rd party one
* Unpinned the specific version of test report action
* Used explicit permissions for the action

### Why are the changes needed?

The problem is the 3rd party action used can't deal with non-zip files. In the orignal [PR](https://github.com/apache/spark/pull/29333) it was mentioned that the official `download-artifact` action should be used when it can download from other workflows. We should try to use that.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

CI

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53167 from gaogaotiantian/fix-report-action.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>