[SPARK-54474][PYTHON] Discard the XML report on tests that are supposed to fail
### What changes were proposed in this pull request?

We do not keep the XML reports from the tests that are supposed to fail anymore.

### Why are the changes needed?

It will confuse XML report consumers. These are tests that are supposed to fail - we are testing if they return the expected return code.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manually tested that no XML report is generated and the test itself still passes.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53187 from gaogaotiantian/disable-xml-on-failed-tests.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>