[SPARK-54247][PYTHON] Explictly close socket for util._load_from_socket
### What changes were proposed in this pull request?

Convert `util._load_from_socket` to a context - so it can close the socket explicitly.

### Why are the changes needed?

The current implementation will generate a warning about socket leak - because it was never closed properly. We rely on gc to close the socket, which is not the end of the world, but we can do better.

### Does this PR introduce _any_ user-facing change?

No, unless the user is using this private util function.

### How was this patch tested?

`pyspark-core` locally, and the rest on CI.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53200 from gaogaotiantian/fix-socket-leak.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>