[SPARK-54450][INFRA] Allow run-tests to accept unittest style string for test names
### What changes were proposed in this pull request?

Allow `run-tests` to take unittest style string like `test_module.TestClass.test_case`, in addition to `test_module TestClass.test_case`.

Also some extra check is added for the existence of the test module - we are not sending it to `bin/pyspark` if we can't find the module.

### Why are the changes needed?

It's extra work to split the string when copy/paste directly from unittest output (CI failures for example).

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manually confirmed the `.` connection works and the original way is not impacted (except that we have some new checks for the module).

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53157 from gaogaotiantian/improve-run-test.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>