[SPARK-54499][SQL][4.1] Re-enable `spark.sql.scripting.enabled` by default for Spark 4.1.0
### What changes were proposed in this pull request?

This PR aims to re-enable SQL Scripting in Spark 4.1, ahead of its RC.

### Why are the changes needed?

There was a previous misunderstanding/miscommunication, that I did not notice in time. The feature was disabled because it was thought that it is not ready, due to unclosed JIRA items.

However, it was just a case of not 100% up-to-date items. I have cleaned up and classified remaining work items (children of [SPARK-48338](https://issues.apache.org/jira/browse/SPARK-48338)) into 4 categories:
- M0 - basic support
- M1 - features and changes required to enable SQL Scripting by default
- M2 - follow-up improvements and additional functionalities that are non-fundamental and should not block M1
- M3 - potential improvements for the future, need investigation

M0 and M1 are done, meaning the feature is stable, useful and also ready to be used.

M2 will improve some of the aspects of using it by introducing a newer and more user-friendly statements, like SIGNAL/RESIGNAL and GET DIAGNOSTICS, as well as do some very minor optimizations. However, without those, feature is still ready to be used as is.

The only missing aspect is documentation, for which we have a PR (https://github.com/apache/spark/pull/50592) but it was closed. I'll work on re-submitting that one.

### Does this PR introduce _any_ user-facing change?

I am not sure 100% how this classifies - the feature was already present in 4.0, but was not enabled by default.
Anyways, the feature is completely orthogonal to all other standalone statement execution paths, so it is just adding to functionality of Spark.
There is no difference in behavior for any standalone SQL statement compared to when SQL Scripting is turned off.

### How was this patch tested?

SQL Scripting has a thorough support in tests.
CI ensures the feature does not affect any other.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53154 from davidm-db/davidm-db/enable-sql-scripting-4.1.

Authored-by: David Milicevic <david.milicevic@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>