[SPARK-54532][PYTHON] Add support for sqlstate for PySparkException
### What changes were proposed in this pull request?

Allow `PySparkException` to get `sqlstate` from the json file.

### Why are the changes needed?

It's useful for python client to throw exceptions with sqlstate.

### Does this PR introduce _any_ user-facing change?

For this specific PR no. It's an infra to have more sqlstate in the future.

### How was this patch tested?

A feature test is added. However, we don't have any sqlstate assigned to python exceptions now - so we can't test the positive case. The future PRs will add more sqlstate and we can test then.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53241 from gaogaotiantian/support-sqlstate.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>