[SPARK-54537][CONNECT] Fix SparkConnectDatabaseMetaData getSchemas/getTables on catalogs missing namespace capbalitity
### What changes were proposed in this pull request?

For DSv2 catalog implementation that does not mix in the interface `SupportsNamespaces`, `SHOW SCHEMAS IN foo_catalog` fails with `MISSING_CATALOG_ABILITY.NAMESPACES`, we should catch this and return empty result instead of failing on `SparkConnectDatabaseMetaData#getSchemas|getTables`

### Why are the changes needed?

Fix a bug in the Connect JDBC dirver.

### Does this PR introduce _any_ user-facing change?

No, Connect JDBC dirver is an unreleased feature.

### How was this patch tested?

UT is adjusted to cover the changes.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53246 from pan3793/SPARK-54537.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>