[SPARK-54533][UI] Set metric ExecutorSource.METRIC_RESULT_SIZE with correct value
### What changes were proposed in this pull request?
Set the metric executorSource.METRIC_RESULT_SIZE after the result size is populated.

### Why are the changes needed?
Currently the metric executorSource.METRIC_RESULT_SIZE is set to the value of `task.metrics.resultSize` which is always 0 in that line.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Manually tested on local Spark. See the Jira attachments for metrics reported before and after fix.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #53243 from jiwen624/executor-source-result-size.

Authored-by: Eric Yang <jiwen624@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>