[SPARK-54572][PYTHON] Support VSCode breakpoints for pyspark
### What changes were proposed in this pull request?

This PR adds a script and a supporting directory for users to have native debugging experience with VSCode (Cursor) for not only driver code, but UDF/workers/daemon.

<img width="2324" height="1010" alt="image" src="https://github.com/user-attachments/assets/57d1bbae-53c6-48e9-8910-25a47e4f0e24" />

### Why are the changes needed?

So user can debug their (local) code with VSCode debugger.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Locally works. This does not touch spark code, it's just a dev tool.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53289 from gaogaotiantian/support-vscode-breakpoint.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>