Revert "[SPARK-54310][SQL] Add `numSourceRows` metric for `MergeIntoExec`"
### What changes were proposed in this pull request?

Clean revert of d65234b8578ce9a01ed656367646e2bea8f0f4fb.

Will later handle for cases of sourceSide child nodes without `numOutputRows`, and will re-target the new implementation to later Spark release.

### Why are the changes needed?

The current implementation may grab the incorrect `numOutputRows` metric if there is an intermediary node (such as custom Spark operator) which does not support the metric.

This is because we target the first sourceSide child node with `numOutputRows`. If a SparkExtension node does not contain this metric but transforms the source table, then we could progress all the way to the source table and grab the incorrect metric.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing CI, as this is a revert

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53293 from asl3/numsourcerowsrevert.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>