[SPARK-54572][PYTHON][FOLLOWUP] Add a file lock for the VSCode magic file
### What changes were proposed in this pull request?

Add a lock to access the VSCode magic file so only one process is allowed to write it and wait for connection from VSCode.

Also we do not track daemon by default because fork could cause issues. We add an option `--hook-daemon` when the user needs to.

### Why are the changes needed?

If multiple processes are writing the file at the same time, VSCode will fail to establish the connection to the debugger server.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manually confirmed that `run-tests` works after the patch.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53318 from gaogaotiantian/fix-vscode-support.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>