[SPARK-54562] Block eager analysis / execution inside flow function from the server side
### What changes were proposed in this pull request?

Eager analysis / execution inside the SDP flow function is not supported. Previously, we enforce such restriction by intercepting the `AnalyzePlan` / `ExecutePlan` RPC  from the Python Spark Connect Client side, which is fragile. It's mostly because during that time, we don't have a good way for the SC server to know whether it is serving a requests that  made inside a flow function.

Now, we have `PipelineAnalysisContext`. If the request comes from the flow function, SC client automatically add a `PipelineAnalysisContext` UserContext extension and SC server can based on its existence to know whether the request comes from flow function. Thereby, we're now able to move the eager analysis / execution validation from the client to the server.

### Why are the changes needed?

Our previous way of performing such validation is very fragile and this PR serves as an code improvement.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New unit tests in PythonPipelineSuite

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53303 from SCHJonathan/jonathan-chang_data/block-eager-analysis.

Authored-by: Yuheng Chang <jonathanyuheng@gmail.com>
Signed-off-by: Sandy Ryza <sandy.ryza@databricks.com>