[SPARK-54628][PYTHON] Remove all unnecessary explicit super() arguments
### What changes were proposed in this pull request?

Remove all unnecessary explicit arguments in super().

```python
class A:
    def __init__(self):
        # from super(A, self).__init__() to
        super().__init__()
````

### Why are the changes needed?

This seems like a cosmetic change but it's more than that. Even though the old syntax works, it has plenty of caveats hence considered bad practice:

1. It's verbose.
2. Most people probably don't know what it means
3. Readers needs to compare the argument with class name to confirm whether it actually changes the MRO order
4. When the class name is changed, it could break
5. When the class is copied, it's very easy to miss the argument and silently skip a method in MRO chain
6. It's very easy to establish [cargo cult effect](https://en.wikipedia.org/wiki/Cargo_cult) where people just use it without knowing why they need to.

There are actually very few explicit usages in our code base that uses a different class than the current class - which is when this explicit convention is supposed to be used. However, I suspect that some of them are actually the fact of neglect from reasons I listed above.

I will take a look at those usages later, but in the meantime, we should use the pure `super()` when it's equivalent.

BTW, this is how python2 works so that's why some people do it, it does not make sense in python3 anymore.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

There should not be any behavior change so wait for the CI

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53369 from gaogaotiantian/remove-super-args.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>