[SPARK-54615][PYTHON] Always pass runner_conf to python worker
### What changes were proposed in this pull request?

Always pass runnerConf to python worker, even if it's not used.

### Why are the changes needed?

This is part of the effort to consolidate our protocol from JVM to the worker. We have different ways to pass the runner conf now and sometimes we just don't pass it. It makes the worker side code a bit messy - we need to determine whether to read the conf based on eval type. However reading an empty conf is super cheap and we can just do it regardless.

With this infra, vanilla python udfs can also pass some runner conf in the future. We can do some refactoring on our JVM worker code in the future.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

`pyspark-sql` passed locally. Running the rest on CI.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53353 from gaogaotiantian/always-pass-runnerconf.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>