[SPARK-54443][SS] Partition key extraction for all streaming stateful operators
### What changes were proposed in this pull request?

We want to extract the same partitioning key, used during shuffle to partition the streaming state, to perform repartition for the operator state. The State store key typically includes the partition key, hence we are extracting it depending on the operator state key schema. Each operator determines how they store data in the state store.

This is to make sure repartition is consistent with how the state will be partitioned during query execution.

### Why are the changes needed?

For offline state repartition

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New tests for each operator

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53355 from micheal-o/partition_key_extract.

Authored-by: micheal-o <micheal.okutubo@gmail.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>