[SPARK-54441][INFRA][FOLLOW-UP] Enforce new forked process to create its own coverage object
### What changes were proposed in this pull request?

Enforce the forked worker process to have it's own coverage object. This is used by the after fork hook for `coveragepy` itself.

### Why are the changes needed?

It seems like some coverage data of workers was not properly recorded. Might be a racing issue. I applied this patch and confirmed that the worker coverage is properly recorded. We should see a better coverage on worker side.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Locally confirmed the new coverage.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53408 from gaogaotiantian/fix-worker-coverage.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>