[SPARK-54644][PYTHON][SS][TESTS] Always stop the query before await it
### What changes were proposed in this pull request?

We have patterns like
```python
        q.processAllAvailable()
        q.awaitTermination(10)
```

which will make the test wait for 10 seconds to timeout - it is such a waste and it's causing our daily coverage run to timeout. We should `stop()` the query before `awaitTermination` and save these 10 seconds - there are a lot of them.

### Why are the changes needed?

It's a waste time and it's causing our CIs to timeout.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Locally it runs much faster (without the unnecessary wait).

### Was this patch authored or co-authored using generative AI tooling?

Closes #53394 from gaogaotiantian/stop-before-await.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>