[SPARK-54653][DOCS] Add cross-session note to cache/persist public APIs
Document that cached data is shared across all Spark sessions within an application for DataFrame/Dataset cache/persist methods and Catalog.cacheTable methods.

### What changes were proposed in this pull request?

This change updates docs and comments to mention about the nature of dataframe cache being cross-session. Only public and newer apis docs are updated, e.g. `SQLContext` could be in scope but since it is a legacy api, no change was made to it.

### Why are the changes needed?

To further clarify the existing behavior for the cache api.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Rebuild the entire project.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53401 from yyanyy/cache-doc-update.

Authored-by: Yan Yan <yyanyyyy@gmail.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>