[SPARK-54632][INFRA][FOLLOW-UP] Enable ruff on our CI and lint-python
### What changes were proposed in this pull request?

This add `ruff` to our docker image and enables ruff check in our CI. Also this adds ruff check in `dev/lint-python`.

We want to have both `ruff` and `flake8` run in CI for a while to confirm the compatibility then we will deprecate `flake8`.

~~It is intentional to leave ruff version blank - so it uses the latest version. I think the linter rule is pretty stable and the version upgrade should not affect our workflow too much. This is an experiment about using the latest version of devtool. Pinning the version has its advantage but upgrading the version is painful. If we can occasionally (per month?) fix a small amount of code for linter, I think that's acceptable.~~

After discussion with Yicong-Huang I think it's a good idea to pin the version - so we can do all the code fix with devtool change in a single PR. We should avoid having multiple people trying to fix the code when there is a new rule.

### Why are the changes needed?

`flake8` is too slow and our pinned version is just too old. We should replace it.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Local ruff test passes.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53412 from gaogaotiantian/add-ruff-to-ci.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>