[SPARK-54609][SQL] Disable TIME type by default
### What changes were proposed in this pull request?

Introducing a new SQL config for TIME type: `spark.sql.timeType.enabled`.

The default value is `false` and it is enabled only in tests.

### Why are the changes needed?

TIME data type support is not complete, so we need to guard it before it is completed, especially ahead of Spark 4.1 release.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Need to add tests for disabled config.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53344 from davidm-db/davidm-db/time-config.

Lead-authored-by: David Milicevic <david.milicevic@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>