[SPARK-54453][PYTHON][TEST] Add more coverage tests to conversion
### What changes were proposed in this pull request?

More coverage tests for `conversion.py`

Also the framework is restructured during this process. It's not easy to add extra test cases to the previous framework because you need to count the number.

Now the schema, input and output are grouped together so anyone can just add new test cases in a single place.

### Why are the changes needed?

To improve coverage and make it easier to add test cases in the future.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

After framework change, confirmed that the coverage is exactly the same as previous. Then confirmed the coverage rate is better after new test cases.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53479 from gaogaotiantian/conversion-test.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>