[SPARK-54711][PYTHON] Add a timeout for daemon created worker connection
### What changes were proposed in this pull request?

A timeout mechanism is introduced for daemon based workers. If the executor can't establish a connection with the spawned worker, it will restart the daemon.

### Why are the changes needed?

Without the mechanism, if the worker somehow hangs before it can establish the connection with the executor, the executor will hang forever.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

It is confirmed to have fixed a daemon worker hanging issue.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53476 from gaogaotiantian/timeout-for-daemon.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Takuya Ueshin <ueshin@databricks.com>