[SPARK-54736][TEST][INFRA] Change the default executable to sys.executable
### What changes were proposed in this pull request?

Use `sys.executable` (the Python that executes the script) instead of `python3.11` for `run-tests.py`.

### Why are the changes needed?

Defaulting to an arbitrary version of python is weird. For example, if I'm running everything in my python3.12 venv, the script will use python3.11 as long as I installed python3.11 on my machine - that's just wrong to me.

We should just use whatever python used for this specific script as the default. Of course when we run `run-tests` (the shell script), it's just `python3`.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Locally works. Need to see if impacts CI.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53510 from gaogaotiantian/change-default-executable.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>