[SPARK-54724][INFRA] Hook worker exception handler to dump report before sending exception
### What changes were proposed in this pull request?

Dump coverage data before the workers send exception to JVM.

### Why are the changes needed?

JVM will kill the workers when it receives an exception from the workers which does not give enough time for the workers to dump coverage data.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Confirmed the exception coverage is missing before applying the patch and exists after.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53495 from gaogaotiantian/fix-coverage-exception.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>